ai-api/pyproject.toml
[project]
name = "flowboard-ai"
version = "0.1.0"
description = "AI services for Flowboard"
requires-python = ">=3.11"

dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "llama-cpp-python",
    "celery",
    "redis",
    "requests",
    "arize-phoenix",
    "weaviate-client>=3.0.0,<4.0.0",
    "sentence-transformers",
    "pydantic-settings",
    "groq"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]